requirements.txt
cdsapi==0.6.1
xarray==2023.8.0
netCDF4==1.6.4
h5netcdf==1.2.0
torch>=2.0.1
torchvision>=0.15.2
numpy>=1.24.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart==0.0.6
pytest==7.4.0
pytest-asyncio==0.21.1
PyYAML==6.0.1